move get_examples
